Moved VirtualNodeCacheTest and VirtualNodeCacheHammerTest to src/test and merged them
