+++ platform-sdk/swirlds-virtualmap/src/test/java/com/swirlds/virtualmap/internal/cache/VirtualNodeCacheTest.java
@@ -48,6 +48,7 @@ import java.nio.charset.StandardCharsets;
 import java.time.Duration;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.Comparator;
 import java.util.HashSet;
 import java.util.Iterator;
@@ -72,7 +73,527 @@ class VirtualNodeCacheTest extends VirtualTestBase {
     // Fast copy and life-cycle tests, including releasing and merging.
     // ----------------------------------------------------------------------
 
-    private final AtomicInteger concurrentTests = new AtomicInteger(0);
+    /**
+     * This is perhaps the most crucial of all the tests here. We are going to build our
+     * test tree, step by step. Initially, there are no nodes. Then we add A, B, C, etc.
+     * We do this in a way that mimics what happens when {@link VirtualMap}
+     * makes the calls. This should be a faithful reproduction of what we will actually see.
+     * <p>
+     * To complicate matters, once we build the tree, we start to tear it down again. We
+     * also do this in order to try to replicate what will actually happen. We make this
+     * even more rich by adding and removing nodes in different orders, so they end up
+     * in different positions.
+     * <p>
+     * We create new caches along the way. We don't drop any of them until the end.
+     */
+    @Test
+    @Tags({@Tag("VirtualMerkle"), @Tag("VirtualNodeCache")})
+    @DisplayName("Build a tree step by step")
+    void buildATree() {
+        // ROUND 0: Add A, B, and C. First add A, then B, then C. When we add C, we have to move A.
+        // This will all happen in a single round. Then create the Root and Left internals after
+        // creating the next round.
+
+        // Add apple at path 1
+        final VirtualNodeCache<TestKey, TestValue> cache0 = cache;
+        final VirtualLeafRecord<TestKey, TestValue> appleLeaf0 = appleLeaf(1);
+        cache0.putLeaf(appleLeaf0);
+        validateLeaves(cache0, 1, Collections.singletonList(appleLeaf0));
+
+        // Add banana at path 2
+        final VirtualLeafRecord<TestKey, TestValue> bananaLeaf0 = bananaLeaf(2);
+        cache0.putLeaf(bananaLeaf0);
+        validateLeaves(cache0, 1, asList(appleLeaf0, bananaLeaf0));
+
+        // Move apple to path 3
+        appleLeaf0.setPath(3);
+        cache0.clearLeafPath(1);
+        cache0.putLeaf(appleLeaf0);
+        assertEquals(DELETED_LEAF_RECORD, cache0.lookupLeafByPath(1, false), "leaf should have been deleted");
+        validateLeaves(cache0, 2, asList(bananaLeaf0, appleLeaf0));
+
+        // Add cherry to path 4
+        final VirtualLeafRecord<TestKey, TestValue> cherryLeaf0 = cherryLeaf(4);
+        cache0.putLeaf(cherryLeaf0);
+        validateLeaves(cache0, 2, asList(bananaLeaf0, appleLeaf0, cherryLeaf0));
+
+        // End the round and create the next round
+        nextRound();
+        validateDirtyLeaves(asList(bananaLeaf0, appleLeaf0, cherryLeaf0), cache0.dirtyLeavesForHash(2, 4));
+
+        // Add an internal node "left" at index 1 and then root at index 0
+        final VirtualHashRecord leftInternal0 = leftInternal();
+        final VirtualHashRecord rootInternal0 = rootInternal();
+        cache0.putHash(leftInternal0);
+        cache0.putHash(rootInternal0);
+        cache0.seal();
+        validateTree(cache0, asList(rootInternal0, leftInternal0, bananaLeaf0, appleLeaf0, cherryLeaf0));
+        final Hash bananaLeaf0intHash = cache0.lookupHashByPath(bananaLeaf0.getPath(), false);
+        assertNull(bananaLeaf0intHash);
+        final Hash appleLeaf0intHash = cache0.lookupHashByPath(appleLeaf0.getPath(), false);
+        assertNull(appleLeaf0intHash);
+        final Hash cherryLeaf0intHash = cache0.lookupHashByPath(cherryLeaf0.getPath(), false);
+        assertNull(cherryLeaf0intHash);
+        // This check (and many similar checks below) is arguable. In real world, dirtyHashes() is only
+        // called when a cache is flushed to disk, and it happens only after VirtualMap copy is hashed, all
+        // hashes are calculated and put to the cache. Here the cache doesn't contain hashes for dirty leaves
+        // (bananaLeaf0, appleLeaf0, cherryLeaf0). Should dirtyHashes() include these leaf nodes? Currently
+        // it doesn't
+        cache0.prepareForFlush();
+        validateDirtyInternals(Set.of(rootInternal0, leftInternal0), cache0.dirtyHashesForFlush(4));
+
+        // ROUND 1: Add D and E.
+        final VirtualNodeCache<TestKey, TestValue> cache1 = cache;
+
+        // Move B to index 5
+        final VirtualLeafRecord<TestKey, TestValue> bananaLeaf1 = bananaLeaf(5);
+        cache1.clearLeafPath(2);
+        cache1.putLeaf(bananaLeaf1);
+        assertEquals(
+                DELETED_LEAF_RECORD,
+                cache1.lookupLeafByPath(2, false),
+                "value that was looked up should match original value");
+        validateLeaves(cache1, 3, asList(appleLeaf0, cherryLeaf0, bananaLeaf1));
+
+        // Add D at index 6
+        final VirtualLeafRecord<TestKey, TestValue> dateLeaf1 = dateLeaf(6);
+        cache1.putLeaf(dateLeaf1);
+        validateLeaves(cache1, 3, asList(appleLeaf0, cherryLeaf0, bananaLeaf1, dateLeaf1));
+
+        // Move A to index 7
+        final VirtualLeafRecord<TestKey, TestValue> appleLeaf1 = appleLeaf(7);
+        cache1.clearLeafPath(3);
+        cache1.putLeaf(appleLeaf1);
+        assertEquals(
+                DELETED_LEAF_RECORD,
+                cache1.lookupLeafByPath(3, false),
+                "value that was looked up should match original value");
+        validateLeaves(cache1, 4, asList(cherryLeaf0, bananaLeaf1, dateLeaf1, appleLeaf1));
+
+        // Add E at index 8
+        final VirtualLeafRecord<TestKey, TestValue> eggplantLeaf1 = eggplantLeaf(8);
+        cache1.putLeaf(eggplantLeaf1);
+        validateLeaves(cache1, 4, asList(cherryLeaf0, bananaLeaf1, dateLeaf1, appleLeaf1, eggplantLeaf1));
+
+        // End the round and create the next round
+        nextRound();
+        validateDirtyLeaves(asList(bananaLeaf1, dateLeaf1, appleLeaf1, eggplantLeaf1), cache1.dirtyLeavesForHash(4, 8));
+
+        // Add an internal node "leftLeft" at index 3 and then "right" at index 2
+        final VirtualHashRecord leftLeftInternal1 = leftLeftInternal();
+        final VirtualHashRecord rightInternal1 = rightInternal();
+        final VirtualHashRecord leftInternal1 = leftInternal();
+        final VirtualHashRecord rootInternal1 = rootInternal();
+        cache1.putHash(leftLeftInternal1);
+        cache1.putHash(rightInternal1);
+        cache1.putHash(leftInternal1);
+        cache1.putHash(rootInternal1);
+        cache1.seal();
+        validateTree(
+                cache1,
+                asList(
+                        rootInternal1,
+                        leftInternal1,
+                        rightInternal1,
+                        leftLeftInternal1,
+                        cherryLeaf0,
+                        bananaLeaf1,
+                        dateLeaf1,
+                        appleLeaf1,
+                        eggplantLeaf1));
+        // prepareForFlush() removes version 0 mutations for paths 2 and 3
+        cache1.prepareForFlush();
+        validateDirtyInternals(
+                Set.of(rootInternal1, leftInternal1, rightInternal1, leftLeftInternal1), cache1.dirtyHashesForFlush(8));
+
+        // ROUND 2: Add F and G
+        final VirtualNodeCache<TestKey, TestValue> cache2 = cache;
+
+        // Move C to index 9
+        final VirtualLeafRecord<TestKey, TestValue> cherryLeaf2 = cherryLeaf(9);
+        cache2.clearLeafPath(4);
+        cache2.putLeaf(cherryLeaf2);
+        assertEquals(
+                DELETED_LEAF_RECORD,
+                cache2.lookupLeafByPath(4, false),
+                "value that was looked up should match original value");
+        validateLeaves(cache2, 5, asList(bananaLeaf1, dateLeaf1, appleLeaf1, eggplantLeaf1, cherryLeaf2));
+
+        // Add F at index 10
+        final VirtualLeafRecord<TestKey, TestValue> figLeaf2 = figLeaf(10);
+        cache2.putLeaf(figLeaf2);
+        validateLeaves(cache2, 5, asList(bananaLeaf1, dateLeaf1, appleLeaf1, eggplantLeaf1, cherryLeaf2, figLeaf2));
+
+        // Move B to index 11
+        final VirtualLeafRecord<TestKey, TestValue> bananaLeaf2 = bananaLeaf(11);
+        cache2.clearLeafPath(5);
+        cache2.putLeaf(bananaLeaf2);
+        assertEquals(
+                DELETED_LEAF_RECORD,
+                cache2.lookupLeafByPath(5, false),
+                "value that was looked up should match original value");
+        validateLeaves(cache2, 6, asList(dateLeaf1, appleLeaf1, eggplantLeaf1, cherryLeaf2, figLeaf2, bananaLeaf2));
+
+        // Add G at index 12
+        final VirtualLeafRecord<TestKey, TestValue> grapeLeaf2 = grapeLeaf(12);
+        cache2.putLeaf(grapeLeaf2);
+        validateLeaves(
+                cache2,
+                6,
+                asList(dateLeaf1, appleLeaf1, eggplantLeaf1, cherryLeaf2, figLeaf2, bananaLeaf2, grapeLeaf2));
+
+        // End the round and create the next round
+        nextRound();
+        validateDirtyLeaves(asList(cherryLeaf2, figLeaf2, bananaLeaf2, grapeLeaf2), cache2.dirtyLeavesForHash(6, 12));
+
+        // Add an internal node "rightLeft" at index 5 and then "leftRight" at index 4
+        final VirtualHashRecord rightLeftInternal2 = rightLeftInternal();
+        final VirtualHashRecord leftRightInternal2 = leftRightInternal();
+        final VirtualHashRecord rightInternal2 = rightInternal();
+        final VirtualHashRecord leftInternal2 = leftInternal();
+        final VirtualHashRecord rootInternal2 = rootInternal();
+        cache2.putHash(rightLeftInternal2);
+        cache2.putHash(leftRightInternal2);
+        cache2.putHash(rightInternal2);
+        cache2.putHash(leftInternal2);
+        cache2.putHash(rootInternal2);
+        cache2.seal();
+        validateTree(
+                cache2,
+                asList(
+                        rootInternal2,
+                        leftInternal2,
+                        rightInternal2,
+                        leftLeftInternal1,
+                        leftRightInternal2,
+                        rightLeftInternal2,
+                        dateLeaf1,
+                        appleLeaf1,
+                        eggplantLeaf1,
+                        cherryLeaf2,
+                        figLeaf2,
+                        bananaLeaf2,
+                        grapeLeaf2));
+        // prepareForFlush() removes version 1 mutations for paths 4 and 5
+        cache2.prepareForFlush();
+        validateDirtyInternals(
+                Set.of(rootInternal2, leftInternal2, rightInternal2, leftRightInternal2, rightLeftInternal2),
+                cache2.dirtyHashesForFlush(12));
+
+        // Now it is time to start mutating the tree. Some leaves will be removed and re-added, some
+        // will be removed and replaced with a new value (same key).
+
+        // Remove A and move G to take its place. Move B to path 5
+        final VirtualNodeCache<TestKey, TestValue> cache3 = cache;
+        final VirtualLeafRecord<TestKey, TestValue> appleLeaf3 = appleLeaf(7);
+        cache3.deleteLeaf(appleLeaf3);
+        assertEquals(
+                DELETED_LEAF_RECORD,
+                cache3.lookupLeafByPath(7, false),
+                "value that was looked up should match original value");
+
+        final VirtualLeafRecord<TestKey, TestValue> grapeLeaf3 = grapeLeaf(7);
+        cache3.clearLeafPath(12);
+        cache3.putLeaf(grapeLeaf3);
+        assertEquals(
+                DELETED_LEAF_RECORD,
+                cache3.lookupLeafByPath(12, false),
+                "value that was looked up should match original value");
+
+        final VirtualLeafRecord<TestKey, TestValue> bananaLeaf3 = bananaLeaf(5);
+        cache3.clearLeafPath(11);
+        cache3.putLeaf(bananaLeaf3);
+        cache3.deleteHash(5);
+        assertEquals(
+                DELETED_LEAF_RECORD,
+                cache3.lookupLeafByPath(11, false),
+                "value that was looked up should match original value");
+        assertEquals(
+                DELETED_HASH,
+                cache3.lookupHashByPath(5, false),
+                "value that was looked up should match original value");
+
+        validateLeaves(cache3, 5, asList(bananaLeaf3, dateLeaf1, grapeLeaf3, eggplantLeaf1, cherryLeaf2, figLeaf2));
+
+        // Add A back. Banana is moved to position 11, Apple goes to position 12
+        appleLeaf3.setPath(12);
+        cache3.putLeaf(appleLeaf3);
+        bananaLeaf3.setPath(11);
+        cache3.putLeaf(bananaLeaf3);
+        cache3.clearLeafPath(5);
+        assertEquals(
+                DELETED_LEAF_RECORD,
+                cache3.lookupLeafByPath(5, false),
+                "value that was looked up should match original value");
+
+        validateLeaves(
+                cache3,
+                6,
+                asList(dateLeaf1, grapeLeaf3, eggplantLeaf1, cherryLeaf2, figLeaf2, bananaLeaf3, appleLeaf3));
+
+        // Update D
+        final VirtualLeafRecord<TestKey, TestValue> dogLeaf3 = dogLeaf(dateLeaf1.getPath());
+        cache3.putLeaf(dogLeaf3);
+
+        // Update F
+        final VirtualLeafRecord<TestKey, TestValue> foxLeaf3 = foxLeaf(figLeaf2.getPath());
+        cache3.putLeaf(foxLeaf3);
+
+        validateLeaves(
+                cache3, 6, asList(dogLeaf3, grapeLeaf3, eggplantLeaf1, cherryLeaf2, foxLeaf3, bananaLeaf3, appleLeaf3));
+
+        // End the round and create the next round
+        nextRound();
+        validateDirtyLeaves(
+                asList(dogLeaf3, grapeLeaf3, foxLeaf3, bananaLeaf3, appleLeaf3), cache3.dirtyLeavesForHash(6, 12));
+
+        // We removed the internal node rightLeftInternal. We need to add it back in.
+        final VirtualHashRecord rightLeftInternal3 = rightLeftInternal();
+        final VirtualHashRecord leftRightInternal3 = leftRightInternal();
+        final VirtualHashRecord leftLeftInternal3 = leftLeftInternal();
+        final VirtualHashRecord rightInternal3 = rightInternal();
+        final VirtualHashRecord leftInternal3 = leftInternal();
+        final VirtualHashRecord rootInternal3 = rootInternal();
+        cache3.putHash(rightLeftInternal3);
+        cache3.putHash(leftRightInternal3);
+        cache3.putHash(leftLeftInternal3);
+        cache3.putHash(rightInternal3);
+        cache3.putHash(leftInternal3);
+        cache3.putHash(rootInternal3);
+        cache3.seal();
+        // prepareForFlush() removes version 1 mutations for paths 6 and 7
+        cache3.prepareForFlush();
+        validateDirtyInternals(
+                Set.of(
+                        rootInternal3,
+                        leftInternal3,
+                        rightInternal3,
+                        leftLeftInternal3,
+                        leftRightInternal3,
+                        rightLeftInternal3),
+                cache3.dirtyHashesForFlush(12));
+
+        // At this point, we have built the tree successfully. Verify one more time that each version of
+        // the cache still sees things the same way it did at the time the copy was made.
+        final VirtualNodeCache<TestKey, TestValue> cache4 = cache;
+        validateTree(
+                cache0,
+                asList(
+                        rootInternal0,
+                        leftInternal0,
+                        null, // became internal in version 1
+                        null, // became internal in version 1
+                        null)); // became internal in version 2
+        validateTree(
+                cache1,
+                asList(
+                        rootInternal1,
+                        leftInternal1,
+                        rightInternal1,
+                        leftLeftInternal1,
+                        null, // became internal in version 1
+                        null, // became internal in version 1
+                        null, // became internal in version 2
+                        null, // became internal in version 2, then updated in version 3
+                        eggplantLeaf1));
+        validateTree(
+                cache2,
+                asList(
+                        rootInternal2,
+                        leftInternal2,
+                        rightInternal2,
+                        leftLeftInternal1,
+                        leftRightInternal2,
+                        rightLeftInternal2,
+                        null, // updated in version 3
+                        null, // updated in version 3
+                        eggplantLeaf1,
+                        cherryLeaf2,
+                        null, // updated in version 3
+                        null, // updated in version 3
+                        null)); // updated in version 3
+        validateTree(
+                cache3,
+                asList(
+                        rootInternal3,
+                        leftInternal3,
+                        rightInternal3,
+                        leftLeftInternal3,
+                        leftRightInternal3,
+                        rightLeftInternal3,
+                        dogLeaf3,
+                        grapeLeaf3,
+                        eggplantLeaf1,
+                        cherryLeaf2,
+                        foxLeaf3,
+                        bananaLeaf3,
+                        appleLeaf3));
+        validateTree(
+                cache4,
+                asList(
+                        rootInternal3,
+                        leftInternal3,
+                        rightInternal3,
+                        leftLeftInternal3,
+                        leftRightInternal3,
+                        rightLeftInternal3,
+                        dogLeaf3,
+                        grapeLeaf3,
+                        eggplantLeaf1,
+                        cherryLeaf2,
+                        foxLeaf3,
+                        bananaLeaf3,
+                        appleLeaf3));
+
+        // Now, we will release the oldest, cache0
+        cache0.release();
+        assertEventuallyDoesNotThrow(
+                () -> {
+                    validateTree(
+                            cache1,
+                            asList(
+                                    rootInternal1,
+                                    leftInternal1,
+                                    rightInternal1,
+                                    leftLeftInternal1,
+                                    null,
+                                    null,
+                                    null,
+                                    null,
+                                    eggplantLeaf1));
+                },
+                Duration.ofSeconds(1),
+                "expected cache1 to eventually become clean");
+        assertEventuallyDoesNotThrow(
+                () -> {
+                    validateTree(
+                            cache2,
+                            asList(
+                                    rootInternal2,
+                                    leftInternal2,
+                                    rightInternal2,
+                                    leftLeftInternal1,
+                                    leftRightInternal2,
+                                    rightLeftInternal2,
+                                    null,
+                                    null,
+                                    eggplantLeaf1,
+                                    cherryLeaf2,
+                                    null,
+                                    null,
+                                    null));
+                },
+                Duration.ofSeconds(1),
+                "expected cache2 to eventually become clean");
+        assertEventuallyDoesNotThrow(
+                () -> {
+                    validateTree(
+                            cache3,
+                            asList(
+                                    rootInternal3,
+                                    leftInternal3,
+                                    rightInternal3,
+                                    leftLeftInternal3,
+                                    leftRightInternal3,
+                                    rightLeftInternal3,
+                                    dogLeaf3,
+                                    grapeLeaf3,
+                                    eggplantLeaf1,
+                                    cherryLeaf2,
+                                    foxLeaf3,
+                                    bananaLeaf3,
+                                    appleLeaf3));
+                },
+                Duration.ofSeconds(1),
+                "expected cache3 to eventually become clean");
+        assertEventuallyDoesNotThrow(
+                () -> {
+                    validateTree(
+                            cache4,
+                            asList(
+                                    rootInternal3,
+                                    leftInternal3,
+                                    rightInternal3,
+                                    leftLeftInternal3,
+                                    leftRightInternal3,
+                                    rightLeftInternal3,
+                                    dogLeaf3,
+                                    grapeLeaf3,
+                                    eggplantLeaf1,
+                                    cherryLeaf2,
+                                    foxLeaf3,
+                                    bananaLeaf3,
+                                    appleLeaf3));
+                },
+                Duration.ofSeconds(1),
+                "expected cache4 to eventually become clean");
+
+        // Now we will release the next oldest, cache 1
+        cache1.release();
+        assertEventuallyDoesNotThrow(
+                () -> {
+                    validateTree(
+                            cache2,
+                            asList(
+                                    rootInternal2,
+                                    leftInternal2,
+                                    rightInternal2,
+                                    null,
+                                    leftRightInternal2,
+                                    rightLeftInternal2,
+                                    null,
+                                    null,
+                                    null,
+                                    cherryLeaf2,
+                                    null,
+                                    null,
+                                    null));
+                },
+                Duration.ofSeconds(1),
+                "expected cache2 to eventually become clean");
+        assertEventuallyDoesNotThrow(
+                () -> {
+                    validateTree(
+                            cache3,
+                            asList(
+                                    rootInternal3,
+                                    leftInternal3,
+                                    rightInternal3,
+                                    leftLeftInternal3,
+                                    leftRightInternal3,
+                                    rightLeftInternal3,
+                                    dogLeaf3,
+                                    grapeLeaf3,
+                                    null, // E hasn't changed since version 1
+                                    cherryLeaf2,
+                                    foxLeaf3,
+                                    bananaLeaf3,
+                                    appleLeaf3));
+                },
+                Duration.ofSeconds(1),
+                "expected cache3 to eventually become clean");
+        assertEventuallyDoesNotThrow(
+                () -> {
+                    validateTree(
+                            cache4,
+                            asList(
+                                    rootInternal3,
+                                    leftInternal3,
+                                    rightInternal3,
+                                    leftLeftInternal3,
+                                    leftRightInternal3,
+                                    rightLeftInternal3,
+                                    dogLeaf3,
+                                    grapeLeaf3,
+                                    null,
+                                    cherryLeaf2,
+                                    foxLeaf3,
+                                    bananaLeaf3,
+                                    appleLeaf3));
+                },
+                Duration.ofSeconds(1),
+                "expected cache to eventually become clean");
+    }
 
     /**
      * Test the public state of a fresh cache. We will test putting, deleting, and clearing
@@ -1981,6 +2502,28 @@ class VirtualNodeCacheTest extends VirtualTestBase {
     // ----------------------------------------------------------------------
     // Bigger Test Scenarios
     // ----------------------------------------------------------------------
+
+    @Test
+    @Tags({@Tag("VirtualMerkle"), @Tag("VirtualNodeCache"), @Tag("DirtyLeaves")})
+    @DisplayName("dirtyLeaves where all mutations are in the same version and none are deleted")
+    void dirtyLeaves_allInSameVersionNoneDeleted() {
+        final VirtualNodeCache<TestKey, TestValue> cache = new VirtualNodeCache<>(VIRTUAL_MAP_CONFIG);
+        cache.putLeaf(appleLeaf(7));
+        cache.putLeaf(bananaLeaf(5));
+        cache.putLeaf(cherryLeaf(4));
+        cache.putLeaf(dateLeaf(6));
+        cache.putLeaf(eggplantLeaf(8));
+        cache.seal();
+
+        final List<VirtualLeafRecord<TestKey, TestValue>> leaves =
+                cache.dirtyLeavesForHash(4, 8).toList();
+        assertEquals(5, leaves.size(), "All leaves should be dirty");
+        assertEquals(cherryLeaf(4), leaves.get(0), "Unexpected leaf");
+        assertEquals(bananaLeaf(5), leaves.get(1), "Unexpected leaf");
+        assertEquals(dateLeaf(6), leaves.get(2), "Unexpected leaf");
+        assertEquals(appleLeaf(7), leaves.get(3), "Unexpected leaf");
+        assertEquals(eggplantLeaf(8), leaves.get(4), "Unexpected leaf");
+    }
 
     @Test
     @Tags({@Tag("VirtualMerkle"), @Tag("VirtualNodeCache"), @Tag("DirtyLeaves")})
